requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python-headless>=4.5.0
numpy>=1.19.0
pandas>=1.3.0
scikit-image>=0.18.0
scikit-learn>=0.24.0
albumentations>=1.0.0
segmentation-models-pytorch>=0.2.0
matplotlib>=3.3.0
tqdm>=4.60.0
pyyaml>=5.4.0
timm>=0.5.0
wandb>=0.12.0  # Optional for experiment tracking
streamlit>=1.10.0
streamlit-image-comparison>=0.0.3
plotly>=5.0.0

